requirements.txt
coverage==7.13.5
django==5.2
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
dotenv==0.9.9
psycopg2-binary==2.9.11